backend/requirements.txt
# Web Framework & Server
fastapi==0.103.2
uvicorn==0.23.2

#database
pymongo==4.7.1
motor==3.3.1

#authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6



# External APIs
spotipy==2.23.0
googlemaps==4.10.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.2
httpx==0.25.0
